[train] Move bin packing and collators into skyrl.train.dataset
Relocate the packing data utilities out of the megatron backend and the
loose top-level module into the skyrl.train.dataset package:

  - skyrl/backends/skyrl_train/distributed/megatron/bin_packing.py
      -> skyrl/train/dataset/bin_packing.py
  - skyrl/train/collators.py
      -> skyrl/train/dataset/collators.py

Pure relocation plus import-path updates, no logic change. The
collators -> bin_packing sibling edge uses a relative import to match
the package's existing intra-package import style; the lazy collators
import in SFTTrainer stays lazy (the collators <-> sft_trainer circular
dependency is unchanged).
